services/presence_bigquery/requirements.txt
google-cloud-error-reporting==1.0.*
google-cloud-bigquery==2.4.*
pymongo==3.11.*